Got 3D rotation working (it seems) by storing location data as floats, and removed _build folder.

+++ point.h
@@ -3,7 +3,7 @@
 #define PI 3.14159265
 
 typedef struct point {
-	int16_t x, y, z;
+	float x, y, z;
 	uint16_t col;
 } point;
 
@@ -13,9 +13,9 @@ void transZ(struct point *p, int16_t dist);
 void rotX(struct point *p, int16_t angle);
 void rotY(struct point *p, int16_t angle);
 void rotZ(struct point *p, int16_t angle);
-void updX(struct point *p, int16_t x);
-void updY(struct point *p, int16_t y);
-void updZ(struct point *p, int16_t z);
+void updX(struct point *p, float x);
+void updY(struct point *p, float y);
+void updZ(struct point *p, float z);
 struct point project(struct point *p, uint16_t fov, uint16_t viewer_dist);
 void draw_pt2D(struct point *p);
 void clear_pt2D(struct point *p);
